+++ tests/drivers/bbram/boards/qemu_riscv32.overlay
@@ -1,0 +1,14 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/ {
+	bbram: bbram@80 {
+		compatible = "ite,it8xxx2-bbram";
+		status = "okay";
+		reg = <0x80 0xff>;
+		label = "BBRAM";
+	};
+};
