+++ tests/drivers/bbram/emul/boards/native_posix.overlay
