+++ tests/drivers/bbram/emul/prj.conf
