+++ tests/drivers/bbram/emul/testcase.yaml
@@ -1,0 +1,10 @@
+# Copyright 2021 Google LLC
+# SPDX-License-Identifier: Apache-2.0
+
+tests:
+  drivers.bbram.emul:
+    tags: drivers bbram
+    harness: ztest
+    platform_allow: native_posix
+    integration_platforms:
+      - native_posix
