+++ tests/drivers/bbram/it8xxx2.conf
@@ -1,0 +1,6 @@
+# Copyright 2021 Google LLC
+# SPDX-License-Identifier: Apache-2.0
+
+CONFIG_ZTEST=y
+CONFIG_BBRAM=y
+CONFIG_BBRAM_IT8XXX2=y
