+++ tests/drivers/bbram/npcx.conf
@@ -1,0 +1,6 @@
+# Copyright 2021 Google LLC
+# SPDX-License-Identifier: Apache-2.0
+
+CONFIG_ZTEST=y
+CONFIG_BBRAM=y
+CONFIG_BBRAM_NPCX=y
