+++ tests/drivers/bbram/testcase.yaml
@@ -1,10 +1,13 @@
 # Copyright 2021 Google LLC
 # SPDX-License-Identifier: Apache-2.0
 
+common:
+  tags: drivers bbram
+  build_only: true
+  harness: ztest
+  platform_exclude: native_posix
 tests:
-  drivers.bbram:
-    tags: drivers bbram
-    harness: ztest
-    platform_allow: native_posix
-    integration_platforms:
-      - native_posix
+  drivers.bbram.it8xxx2:
+    extra_args: CONF_FILE="it8xxx2.conf"
+  drivers.bbram.npcx:
+    extra_args: CONF_FILE="npcx.conf"
